+++ src/Adapter/Create/Create.mjs
@@ -1,3 +1,3 @@
 /**
- * @typedef {{confirm_password: string, password: string, semester: string}} Create
+ * @typedef {{"confirm-password": string, password: string, semester: string}} Create
  */