+++ src/Adapter/Start/Start.mjs
@@ -1,5 +1,5 @@
 /** @typedef {import("../Semester/Semester.mjs").Semester} Semester */
 
 /**
- * @typedef {{minPasswordLength: number, semesters: Semester[]}} Start
+ * @typedef {{"min-password-length": number, semesters: Semester[]}} Start
  */