+++ src/Adapter/Subject/Subject.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {{id: string, label: string, qualifications: {id: string, label: string, required?: boolean}[]}} Subject
- */